Separate ConstantBuffer to API-independent and -specific implementations, and unify the header. Add dummy D3D9 constant buffer implementation (that only logs error if used) so that user will not get link errors if attempting to use.
