+++ Source/Urho3D/Graphics/Direct3D9/D3D9ConstantBuffer.cpp
@@ -23,7 +23,6 @@
 #include "../../Precompiled.h"
 
 #include "../../Graphics/Graphics.h"
-#include "../../Graphics/GraphicsImpl.h"
 #include "../../Graphics/ConstantBuffer.h"
 #include "../../IO/Log.h"
 
@@ -34,65 +33,20 @@ namespace Urho3D
 
 void ConstantBuffer::OnDeviceReset()
 {
-    // No-op on Direct3D11
 }
 
 void ConstantBuffer::Release()
 {
-    URHO3D_SAFE_RELEASE(object_.ptr_);
-
-    shadowData_.Reset();
-    size_ = 0;
 }
 
 bool ConstantBuffer::SetSize(unsigned size)
 {
-    Release();
-
-    if (!size)
-    {
-        URHO3D_LOGERROR("Can not create zero-sized constant buffer");
-        return false;
-    }
-
-    // Round up to next 16 bytes
-    size += 15;
-    size &= 0xfffffff0;
-
-    size_ = size;
-    dirty_ = false;
-    shadowData_ = new unsigned char[size_];
-    memset(shadowData_.Get(), 0, size_);
-
-    if (graphics_)
-    {
-        D3D11_BUFFER_DESC bufferDesc;
-        memset(&bufferDesc, 0, sizeof bufferDesc);
-
-        bufferDesc.ByteWidth = size_;
-        bufferDesc.BindFlags = D3D11_BIND_CONSTANT_BUFFER;
-        bufferDesc.CPUAccessFlags = 0;
-        bufferDesc.Usage = D3D11_USAGE_DEFAULT;
-
-        HRESULT hr = graphics_->GetImpl()->GetDevice()->CreateBuffer(&bufferDesc, 0, (ID3D11Buffer**)&object_.ptr_);
-        if (FAILED(hr))
-        {
-            URHO3D_SAFE_RELEASE(object_.ptr_);
-            URHO3D_LOGD3DERROR("Failed to create constant buffer", hr);
-            return false;
-        }
-    }
-
-    return true;
+    URHO3D_LOGERROR("Constant buffers are not supported on Direct3D9");
+    return false;
 }
 
 void ConstantBuffer::Apply()
 {
-    if (dirty_ && object_.ptr_)
-    {
-        graphics_->GetImpl()->GetDeviceContext()->UpdateSubresource((ID3D11Buffer*)object_.ptr_, 0, 0, shadowData_.Get(), 0, 0);
-        dirty_ = false;
-    }
 }
 
 }
